schedule class code

+++ src/main/java/com/msu/repositories/CourseSemesterMappingRepository.java
@@ -11,6 +11,6 @@ public interface CourseSemesterMappingRepository extends JpaRepository<CourseSem
 
 	
 	 public List<CourseSemesterMapping> findBySemesterId(Integer semesterId);
-	 public CourseSemesterMapping findByCourseSemesterMappingId(Integer courseSemesterMappingId);
+	 public CourseSemesterMapping findByCourseSemesterMappingId(Long courseSemesterMappingId);
 	 
 }
